checkNodePortsAndNotify added
